requirements.txt
altair
asttokens 
attrs 
blinker 
Brotli 
cachetools 
certifi 
cffi 
charset-normalizer 
click 
colorama 
comm 
debugpy 
decorator 
duckdb==1.0.0
entrypoints 
et-xmlfile 
exceptiongroup
executing
fastjsonschema
gitdb 
GitPython 
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gspread==5.12.4
gspread-dataframe==4.0.0
gspread-formatting==1.2.0
gspread-pandas==3.3.0
h2
hpack==4.0.0
httplib2==0.22.0
hyperframe 
idna 
importlib_metadata 
importlib_resources 
ipykernel 
ipython 
ipywidgets 
jedi 
Jinja2 
jsonschema 
jsonschema-specifications 
jupyter_client 
jupyter_core 
jupyterlab_widgets
lxml
markdown-it-py 
MarkupSafe 
matplotlib-inline 
mdurl 
narwhals 
nbformat 
nest_asyncio 
millify
numpy
oauth2client
oauthlib==3.2.2
openpyxl 
packaging 
pandas 
parso 
pickleshare
pillow 
pkgutil_resolve_name 
platformdirs 
plotly 
polars 
prompt_toolkit 
proto-plus==1.24.0
protobuf==4.25.3
psutil 
pure_eval 
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser 
pydeck 
Pygments 
pyparsing==3.1.2
PySocks 
python-dateutil 
pytz 
# pywin32==306
PyYAML 
pyzmq 
referencing 
requests 
requests-oauthlib==2.0.0
rich 
rpds-py 
rsa==4.9
setuptools==72.1.0
six 
smmap 
sql_metadata==2.12.0
sqlparse==0.5.1
statsmodels
st-gsheets-connection==0.1.0
stack-data 
streamlit 
tenacity 
toml 
tornado 
traitlets 
typing_extensions 
tzdata 
tzlocal 
uritemplate==4.1.1
urllib3 
validators 
watchdog 
wcwidth 
wheel==0.44.0
widgetsnbextension
win-inet-pton
zipp 
zstandard==0.23.0
